Crate glue

This crate adds networking functionality and builds upon the functions provided in ZeroMQ.

Modules

pipeline

publisher 🔒

subscriber 🔒

sync

types

All public facing glue *Msg types.

util

This module provides abstractions for serializing and deserializing data into various formats, specifically JSON strings and byte arrays.

Statics

ZMQ_CONTEXT

Traits

GluePublisher

GlueSubscriber